refractor: move aframe files out of components subfolder
Aframe files only define functions - they are not components themselves
