+++ src/Aframe/render-2d-clipplane.js
